requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
aiosqlite==0.19.0
python-dotenv==1.0.0
pandas
requests
click
ixnetwork_restpy